requirements.txt
streamlit==1.49.0
python-dotenv==1.0.1
loguru==0.7.3
pydantic==2.12.5
openai==1.66.3